+++ src/example-archive/simple-examples/working/add_4.c
@@ -2,9 +2,9 @@
 
 signed int inc_1(signed int i)
 /*@ requires 
-      let MAXi32 = 2147483647i64; 
-      (i64) i + 1i64 <  MAXi32;
-    ensures return == i + 1i32; @*/
+      let MAXi32 = 2147483647; 
+      i + 1 <  MAXi32;
+    ensures return == i + 1; @*/
 {
   i = i + 1;
   return i;
