+++ src/example-archive/simple-examples/working/inc_1.c
@@ -1,9 +1,9 @@
 
 int inc_1_pre(int i) 
 /*@ requires 
-      let MAXi32 = 2147483647i64; 
-      (i64) i + 1i64 <  MAXi32;
-    ensures return == i + 1i32; @*/
+      let MAXi32 = 2147483647; 
+      i + 1 <  MAXi32;
+    ensures return == i + 1; @*/
 { 
   int start, pre, post; 
   start = i; 
@@ -14,9 +14,9 @@ int inc_1_pre(int i)
 
 int inc_1_post(int i) 
 /*@ requires 
-      let MAXi32 = 2147483647i64; 
-      (i64) i + 1i64 <  MAXi32;
-    ensures return == i + 1i32; @*/
+      let MAXi32 = 2147483647; 
+      i + 1 <  MAXi32;
+    ensures return == i + 1; @*/
 { 
   int start, pre, post; 
   start = i;
