+++ src/example-archive/simple-examples/working/loop_3.c
@@ -5,9 +5,9 @@
 
 int loop_3(int i)
 /*@ requires 
-      let MAXi32 = (i64) 2147483647i64; // TODO: lift to library 
-      (i64) i + 1i64 <  MAXi32; 
-      0i32 < i; @*/
+      let MAXi32 = 2147483647; // TODO: lift to library 
+      i + 1 <  MAXi32; 
+      0 < i; @*/
 // TODO: ensures? 
 {
   int n = 0;
@@ -15,7 +15,7 @@ int loop_3(int i)
 
   while (n != i)
   /*@ inv n <= i; 
-          0i32 <= acc; 
+          0 <= acc; 
           acc <= n; @*/
   {
     acc = n - acc;
